+++ grakn-graql/src/main/java/ai/grakn/graql/internal/reasoner/atom/binary/TypeAtom.java
@@ -89,7 +89,7 @@ public abstract class TypeAtom extends Binary{
     public boolean isSelectable() {
         return getTypePredicate() == null
                 //disjoint atom
-                || !this.getNeighbours().findFirst().isPresent()
+                || !this.getNeighbours(Atom.class).findFirst().isPresent()
                 || isRuleResolvable();
     }
